+++ src/remotes/dzrpbuffer/dzrpbufferremote.ts
@@ -552,7 +552,7 @@ export class DzrpBufferRemote extends DzrpRemote {
 	 */
 	protected async sendDzrpCmdAddBreakpoint(bpAddress: number, condition?: string): Promise<number> {
 		// Break because 'Continue' works with a temporary breakpoint list.
-		// TODO: Restart after break.
+		// TODO: I should change this when I remove CM_ADD_BREAKPOINT. If the system is running then the one additional breakpoint is sent. If not then nothing happens.
 		await this.sendDzrpCmd(DZRP.CMD_PAUSE);
 		// Convert condition string to Buffer
 		if (!condition)
@@ -570,7 +570,7 @@ export class DzrpBufferRemote extends DzrpRemote {
 	 */
 	protected async sendDzrpCmdRemoveBreakpoint(bpId: number): Promise<void> {
 		// Break because 'Continue' works with a temporary breakpoint list.
-		// TODO: Restart after break.
+		// TODO: I should change this when I remove CMD_REMOVE_BREAKPOINT. If the system is running then the one additional breakpoint is removed. If not then nothing happens.
 		await this.sendDzrpCmd(DZRP.CMD_PAUSE);
 		await this.sendDzrpCmd(DZRP.CMD_REMOVE_BREAKPOINT, [bpId&0xFF, bpId>>>8]);
 	}