+++ src/remotes/dzrp/dzrpremote.ts
@@ -18,7 +18,7 @@ import {Log} from '../../log';
 
 
 // The current implemented version of the protocol.
-export const DZRP_VERSION=[1, 4, 0];
+export const DZRP_VERSION=[1, 6, 0];
 
 // The program name and version transmitted during CMD_INIT.
 export const DZRP_PROGRAM_NAME="DeZog v"+process.version;
@@ -31,25 +31,28 @@ export const DZRP_PROGRAM_NAME="DeZog v"+process.version;
 export enum DZRP {
 	// ZXNext: All Commands available in ZXNext (need to be consecutive)
 	CMD_INIT=1,
-	CMD_GET_REGISTERS=2,
-	CMD_SET_REGISTER=3,
-	CMD_WRITE_BANK=4,
-	CMD_CONTINUE=5,
-	CMD_PAUSE=6,
-	CMD_READ_MEM=7,
-	CMD_WRITE_MEM=8,
-	CMD_GET_SLOTS=9,
-	CMD_SET_SLOT=10,
-	CMD_GET_TBBLUE_REG=11,
-	CMD_SET_BORDER=12,
-	CMD_SET_BREAKPOINTS=13,
-	CMD_RESTORE_MEM=14,
-	CMD_GET_SPRITES_PALETTE=15,
-	CMD_GET_SPRITES_CLIP_WINDOW_AND_CONTROL=16,
+
+	CMD_CLOSE=2,
+	CMD_GET_REGISTERS=3,
+	CMD_SET_REGISTER=4,
+	CMD_WRITE_BANK=5,
+	CMD_CONTINUE=6,
+	CMD_PAUSE=7,
+	CMD_READ_MEM=8,
+	CMD_WRITE_MEM=9,
+	CMD_GET_SLOTS=10,
+	CMD_SET_SLOT=11,
+	CMD_GET_TBBLUE_REG=12,
+	CMD_SET_BORDER=13,
+	CMD_SET_BREAKPOINTS=14,
+	CMD_RESTORE_MEM=15,
+	CMD_LOOPBACK=16,
+	CMD_GET_SPRITES_PALETTE=17,
+	CMD_GET_SPRITES_CLIP_WINDOW_AND_CONTROL=18,
 
 	// Sprites
-	CMD_GET_SPRITES=30,
-	CMD_GET_SPRITE_PATTERNS=31,
+	CMD_GET_SPRITES=19,
+	CMD_GET_SPRITE_PATTERNS=20,
 
 	// Breakpoint
 	CMD_ADD_BREAKPOINT=40,
@@ -1245,7 +1248,16 @@ export class DzrpRemote extends RemoteBase {
 	 */
 	protected async sendDzrpCmdInit(): Promise<{error: string|undefined, programName: string, dzrpVersion: string}> {
 		Utility.assert(false);
-		return {error: undefined, dzrpVersion:"", programName: ""};
+		return {error: undefined, dzrpVersion: "", programName: ""};
+	}
+
+
+	/**
+	 * Override.
+	 * The last command sent. Closes the debug session.
+	 */
+	protected async sendDzrpCmdClose(): Promise<void> {
+		Utility.assert(false);
 	}