+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # 1.4.2
 - Support for DZRP 1.4.0: Changed numbering.
 - ZXNext remote debugging working. Major changes to DZRP.
+- Breakpoints at address 0x0000 fixed.
 
 # 1.4.1
 - Support for DZRP 1.2.0: CMD_SLOT.